% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{addDateOfReportRun}
\alias{addDateOfReportRun}
\title{Add date of report running}
\usage{
addDateOfReportRun(pathsInfo)
}
\arguments{
\item{pathsInfo}{matrix, see output from \code{\link{getMetadata}}.}
}
\value{
A matrix, same as input \code{pathsInfo} with an extra row with the date 
of today.
}
\description{
Add the today's date of when the report runs to the info of the metadata.
}
