% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alertData.R
\name{alertData}
\alias{alertData}
\title{Create alert data}
\usage{
alertData(
  data,
  dataPath = ".",
  alerts,
  subjectVar = "USUBJID",
  verbose = FALSE,
  labelVars = NULL,
  labelData = "data"
)
}
\arguments{
\item{data}{Data.frame with data to create the alerts.}

\item{dataPath}{String with path to the data.}

\item{subjectVar}{String with subject ID variable, 'USUBJID' by default.}

\item{verbose}{Logical, if TRUE (FALSE by default) progress messages are printed
in the current console.
For the visualizations, progress messages during download
of subject-specific report are displayed in the browser console.}

\item{labelVars}{Named character vector containing variable labels of \code{data}.
This will be updated with the labels of the extra annotation variables
(in \code{attr(output, 'labelVars')}).}

\item{labelData}{(optional) String with label for input \code{data},
that will be included in progress messages.}
}
\description{
Create alert data
}
\author{
Michela Pasetto
}
