% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateData.R
\name{annotateData}
\alias{annotateData}
\title{Annotate a SDTM dataset.}
\usage{
annotateData(data, dataPath = ".", annotations, subjectVar = "USUBJID")
}
\arguments{
\item{data}{Data.frame with input data to annotate.}

\item{dataPath}{String with path to the data.}

\item{annotations}{Annotations (or list of those) either as a:
\itemize{
\item{string with standard annotation type, among:
\itemize{
\item{demographics: }{standard variables from the DM data are extracted}
\item{exposed_subjects: }{a logical variable: \code{EXFL} is added to \code{data},
identifying exposed subjects, i.e. subjects included in the 'EX'
dataset and with non empty and non missing: 'EXSTDTC' are considered}
}
}
\item{list of custom annotation, with:
\itemize{
\item{'dataset' (option 1): }{String with name of the annotation dataset,
 e.g. 'ex' to import data from the file: '[dataset].sas7bdat'in \code{dataPath}}
\item{'data' (option 2): }{Data.frame with annotation dataset}
\item{'vars': }{Character vector with variables of interest from annotation dataset.
If not specified, all variables of the dataset are considered.}
\item{'filters': }{Filters for the annotation dataset, 
see \code{filters} parameter of \code{\link{filterData}}
}
}
}
}}

\item{subjectVar}{String with subject ID variable, 'USUBJID' by default.}
}
\value{
Annotated \code{data}
}
\description{
Standard annotation variables are available via the 
parameter \code{annotType}. Custom dataset/variables of interest
are specified via the \code{annotDataset}/\code{annotVar} parameters.
}
\examples{
library(glpgUtilityFct)

data(SDTMDataPelican)

dataLB <- SDTMDataPelican$LB
dataDM <- SDTMDataPelican$DM
dataEX <- SDTMDataPelican$EX

# standard annotations:
# path to dataset should be specified via: 'pathData'
\dontrun{
annotateData(dataLB, annotations = "demographics", pathData = ...)
}

# add all variables in annotation data (if not already available)
head(annotateData(dataLB, annotations = list(data = dataDM)), 1)

# only variables of interest
head(annotateData(dataLB, annotations = list(data = dataDM, vars = c("ARMCD", "COUNTRY"))), 1)

# filter annotation dataset
dataAnnotated <- annotateData(dataLB, 
	annotations = list(
		data = dataDM, 
		vars = c("ARMCD", "COUNTRY", "SEX"), 
		filters = list(var = "SEX", value = "F"))
	)
head(subset(dataAnnotated, SEX == "F"), 1)
head(subset(dataAnnotated, is.na(SEX)), 1)

# multiple annotations:
dataAnnotated <- annotateData(dataLB, 
	annotations = list(
		list(data = dataDM, vars = c("ARMCD", "COUNTRY", "SEX")),
		list(data = dataEX)
	)
)
head(dataAnnotated, 1)
head(subset(dataAnnotated, !is.na(EXSEQ)), 1)
}
