% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-metadata.R
\name{checkAvailabilityMetadata}
\alias{checkAvailabilityMetadata}
\title{Check availability of arguments in list}
\usage{
checkAvailabilityMetadata(paramsList, subListName)
}
\arguments{
\item{paramsList}{A named list.}

\item{subListName}{String indicating which of the sublist names to check for existance.}
}
\value{
The content of the sublist. If not available, returns "Not Available".
}
\description{
Check availability of arguments in list
}
