% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-templates.R
\name{checkConfigFile}
\alias{checkConfigFile}
\title{Check a configuration file (in _YAML_ format)
based on a requirement file in JSON Schema format.}
\usage{
checkConfigFile(configFile, configSpecFile)
}
\arguments{
\item{configFile}{String with name of the config
file of interest in YAML format.}

\item{configSpecFile}{String with name of the config file
containing requirements in JSON Schema format.}
}
\value{
No returned value, an error message is printed
in the console if the configuration file doesn't comply
to the specified specifications.
}
\description{
Check a configuration file (in _YAML_ format)
based on a requirement file in JSON Schema format.
}
\author{
Laure Cougnaud
}
