% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_medicalMonitoringReport.R
\name{checkTemplatesName}
\alias{checkTemplatesName}
\title{Checks of config files template.}
\usage{
checkTemplatesName(configFiles, configDir = "./config")
}
\arguments{
\item{configFiles}{Character vector with name or path of the config file(s).}

\item{configDir}{String with directory with config files,
('config' by default).
It should contain a general 'config.yml' file and dedicated
'config-[X].yml' for each chapter.
The order of each chapter is specified in the 'config' slot in the general 
general 'config.yml'.}
}
\value{
Updated \code{configFiles}
}
\description{
Check if the templates specified in the input config files
don't originate from multiple sources (e.g. custom and R package
via the parameter \code{templatePackage}).
If so, the corresponding config files are not considered.
}
\author{
Laure Cougnaud
}
