% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly-utility-checkValueType.R
\name{checkValueType}
\alias{checkValueType}
\title{Check if the specified \code{valueType} parameter can be passed to the
\code{branchvalues} of the \code{\link[plotly]{plot_ly}} treemap/sunburst visualizations.}
\usage{
checkValueType(
  data,
  vars,
  varsLab = getLabelVar(vars, labelVars = labelVars),
  valueVar,
  valueType = "total",
  labelVars = NULL
)
}
\arguments{
\item{data}{Data.frame with data.}

\item{labelVars}{Named character vector containing variable labels,
used by default for all labels in the plot.}
}
\value{
If the condition is fullfilled:
updated \code{valueType} and warning;
otherwise input \code{valueType}.
}
\description{
If this parameter is set to 'total' and the sum of the counts of the
the children nodes is not bigger than the parent node,
an empty plot is created. In this case, this function
set this parameter to: 'relative'.
}
\author{
Laure Cougnaud
}
