% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTables.R
\name{compareTables}
\alias{compareTables}
\title{Compare tables}
\usage{
compareTables(
  newData,
  oldData,
  referenceVars = NULL,
  changeableVars = NULL,
  labelVars = NULL,
  patientProfilePath = NULL
)
}
\arguments{
\item{newData}{data.frame object representing the new data}

\item{oldData}{data.frame object representing the old data}

\item{referenceVars}{character vector of the columns in the data that are the used as reference for the comparison.
The \code{referenceVars} need to have the same name in both the new and old data sets.}

\item{changeableVars}{character vector of the columns in the data for which you want to assess the change,
e.g. variables that might have changed from the old to the new data.
The \code{changeableVars} need to have the same name in both the new and old data sets.}

\item{labelVars}{named character vector for renaming the columns of the output table.
If NULL, the column names of the data frame are used.}
}
\value{
Return an interactive table showing the changes.
Additions and removals are green or red colored, respectively.
}
\description{
Compare tables (data frames) by showing rows that have been added or removed or changed.
}
\author{
Michela Pasetto
}
