% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_medicalMonitoringReport.R
\name{convertMdToHtml}
\alias{convertMdToHtml}
\title{Convert medical monitoring Markdown files to HTML}
\usage{
convertMdToHtml(
  outputDir = "./MOMP",
  intermediateDir = "./interim",
  configDir = "./config",
  mdFiles = NULL,
  indexPath = "index.Rmd"
)
}
\arguments{
\item{outputDir}{String with output directory,
('MOMP', for: 'Medical Oversight and monitoring Plan' by default).}

\item{intermediateDir}{String with intermediate directory ('inter'
by default), where
markdown files and rds file specifying Js libraries (with \code{knit_meta}) for
each sub report are stored.}

\item{configDir}{String with directory with config files,
('config' by default).
It should contain a general 'config.yml' file and dedicated
'config-[X].yml' for each chapter.
The order of each chapter is specified in the 'config' slot in the general 
general 'config.yml'.}

\item{mdFiles}{(optional) Path to the \code{Markdown} files that
should be converted. If specified, the specified config files 
in \code{configDir} are ignored.}

\item{indexPath}{String with path to the index file,
('index.Rmd' by default).}
}
\value{
String with path to the front page of the 
medical monitoring report.
}
\description{
This consists of:
\enumerate{
\item{importing the general config file ('config'.yml) to identify
each report of interest ('config' tag)}
\item{for each report of interest:
checking if the associated \code{Markdown} and \code{rds} file
(list of Js dependencies) are available in \code{intermediateDir}}
\item{combining all \code{Rmarkdown} reports to a single document: \code{main.md}}
\item{converting \code{main.md} to an HTML document}
}
}
\seealso{
Other medical monitoring reporting: 
\code{\link{getMdFromConfig}()},
\code{\link{getMdHeader}()},
\code{\link{getParamsFromConfig}()},
\code{\link{gitbook_medicalMonitoring_report}()},
\code{\link{html_medicalMonitoring_report}()},
\code{\link{knitPrintMedicalMonitoring}()},
\code{\link{render_medicalMonitoringReport}()}
}
\author{
Laure Cougnaud
}
\concept{medical monitoring reporting}
