% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{createPatientProfileVar}
\alias{createPatientProfileVar}
\title{Create link to patient profile}
\usage{
createPatientProfileVar(data, patientProfilePath, subjectVar = "USUBJID")
}
\arguments{
\item{data}{a data.frame}

\item{subjectVar}{string indicating which column in the data represents the
unique subject identifier, "USUBJID" by default.}

\item{character}{indicating the path to the patient profile directory}
}
\value{
A data.frame with two extra columns:
\code{patientProfilePath} and \code{patientProfileLink} with
the path to the directory and the path to each patient profile, respectively.
}
\description{
Create a link to a patient profile directory
(where the patient profile files are saved) by adding an extra column with the link
in the data.
}
\author{
Michela Pasetto
}
