% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipMedicalMonitoring.R
\name{createRedirectPage}
\alias{createRedirectPage}
\title{Create a redirect page}
\usage{
createRedirectPage(redirectPage = "report.html", dir = "./report_dependencies")
}
\arguments{
\item{redirectPage}{String with the path of the html file that redirects
to the "1-introduction.html" page of the report.}

\item{dir}{String for the path where the "1-introduction.html" is stored.}
}
\value{
The html file is created.
}
\description{
Create an html page that redirects to the 
"1-introduction.html" page of the medical monitoring report available in 
a directory.
See output from \code{\link{render_medicalMonitoring}}.
}
