% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-templates.R
\name{createTemplateDoc}
\alias{createTemplateDoc}
\title{Create documentation for medical monitoring template reports
available in the 'template' folder of the package.}
\usage{
createTemplateDoc()
}
\value{
Character vector with Rd code containing description
for all template documents.
}
\description{
If a JSON schema file available, the information relative
to the template is extracted from this file with the function
\code{JSONSchToRd}.
}
\references{
\href{JSON schema specification}{https://json-schema.org/understanding-json-schema/}
}
\author{
Laure Cougnaud
}
