% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTables.R
\name{formatComparisonDF}
\alias{formatComparisonDF}
\title{Format comparison table}
\usage{
formatComparisonDF(comparisonDF, labelVars, patientProfilePath)
}
\arguments{
\item{comparisonDF}{data.frame object output from getComparisonDF}

\item{labelVars}{named character vector for renaming the columns of the output table.
If NULL, the column names of the data frame are used.}
}
\value{
A data frame formated to be used for the interactive table in \code{compareTables}.
The 'Type' column indicates what kind of change has occured in the data whereas
the 'Version' column indicates whether the change happens in the old (Previous) or new (Current) data.
}
\description{
Format a comparison table (data frame) by adding 'Type' and 'Version' columns.
}
\author{
Michela Pasetto
}
