% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-formatData.R
\name{formatDataForPlotMonitoring}
\alias{formatDataForPlotMonitoring}
\title{Format data for interactive plot for medical monitoring.}
\usage{
formatDataForPlotMonitoring(
  data,
  hoverVar = NULL,
  hoverLab = getLabelVar(hoverVar, labelVars = labelVars),
  hoverByVar = NULL,
  keyVar = NULL,
  id = paste0("plotMonitoring", sample.int(n = 1000, size = 1)),
  labelVars = NULL
)
}
\arguments{
\item{data}{Data.frame with data.}

\item{hoverVar}{Character vector with variables to be displayed in the hover,
by default \code{xVar}, \code{yVar} and any aesthetic variables.}

\item{hoverLab}{Named character vector with labels for \code{hoverVar}.}

\item{hoverByVar}{Character vector with variables identifying
unique elements in the plot, usually x, y, facet variables.
These variables are used to identify records with the same position
in the plot, their information are combined in the hover.}

\item{keyVar}{Character vector with key variables, identifying unique
group for which the link between the table and the plot should be done.}

\item{id}{String with general id for the plot:
\itemize{
\item{'id' is used as \code{group} for the \code{\link[crosstalk]{SharedData}}}
\item{'button:[id]' is used as button ID if \code{table} is TRUE}
}
If not specified, a random id, as 'plotMonitoring[X]' is used.}

\item{labelVars}{Named character vector containing variable labels,
used by default for all labels in the plot.}
}
\description{
Format data for interactive plot for medical monitoring.
}
\author{
Laure Cougnaud
}
