% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{formatPathDateInfoMetadata}
\alias{formatPathDateInfoMetadata}
\title{Format the info on paths from metadata}
\usage{
formatPathDateInfoMetadata(pathsInfo)
}
\arguments{
\item{pathsInfo}{matrix, see output from \code{\link{getMetadata}}.}
}
\value{
A kable object, to be printed.
}
\description{
Format the info on paths from metadata
}
