% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-formatPlotlyMonitoring.R
\name{formatPlotlyMonitoring}
\alias{formatPlotlyMonitoring}
\title{Format interactive plot,
with possibility to download patient profiles
on a click event.}
\usage{
formatPlotlyMonitoring(
  pl,
  data,
  idVar = "USUBJID",
  pathVar = NULL,
  highlightOn = "plotly_click",
  highlightOff = "plotly_doubleclick"
)
}
\arguments{
\item{pl}{\code{\link[plotly]{ggplotly}} object.}

\item{highlightOn}{String with event to turn on the selection
(\code{on} parameter of \code{\link[plotly]{highlight}}),
'plotly_click' by default.}

\item{highlightOff}{String with event to turn off the selection
(\code{off} parameter of \code{\link[plotly]{highlight}}),
'plotly_doubleclick' by default.}
}
\value{
Updated \code{\link[plotly]{ggplotly}} object.
}
\description{
Format interactive plot,
with possibility to download patient profiles
on a click event.
}
\author{
Laure Cougnaud
}
