% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly-utility-formatToHierarchicalData.R
\name{formatToHierarchicalData}
\alias{formatToHierarchicalData}
\title{Format data to a hierarchical data, in the format
as required by the plotly sunburst and treemap.}
\usage{
formatToHierarchicalData(data, vars)
}
\arguments{
\item{data}{Data.frame with data.}

\item{vars}{Character vector with variable(s) of interest.}
}
\value{
Updated data.frame with \code{vars} in 
hierarchical format.
}
\description{
Format data to a hierarchical data, in the format
as required by the plotly sunburst and treemap.
}
\author{
Laure Cougnaud
}
