% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTables.R
\name{getColumnsIdx}
\alias{getColumnsIdx}
\title{Extract position indices of columns in a data frame}
\usage{
getColumnsIdx(df, vars)
}
\arguments{
\item{df}{data frame}

\item{vars}{character vector of columns of the data frame}
}
\description{
Extract position indices of columns in a data frame
}
\author{
Michela Pasetto
}
