% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTables.R
\name{getComparisonDF}
\alias{getComparisonDF}
\title{Get comparison table}
\usage{
getComparisonDF(newData, oldData, referenceVars)
}
\arguments{
\item{newData}{data.frame object representing the new data}

\item{oldData}{data.frame object representing the old data}

\item{referenceVars}{character vector of the columns in the data that are the used as reference for the comparison.
The \code{referenceVars} need to have the same name in both the new and old data sets.}
}
\value{
A data frame from binding the \code{comparison_df} and \code{comparison_table_diff} from the \code{compareDF} package.
}
\description{
Extract a comparison table (data frame) by showing rows that have been added or removed or changed.
}
\author{
Michela Pasetto
}
