% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-referenceLines.R
\name{getFacetVars}
\alias{getFacetVars}
\title{Get facetting variables from facettin parameters.}
\usage{
getFacetVars(facetPars = NULL)
}
\arguments{
\item{facetPars}{List with facetting parameters, passed to the facetting function.}
}
\value{
Character vector with facetting variable
}
\description{
Get facetting variables from facettin parameters.
}
\author{
Laure Cougnaud
}
