% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTables.R
\name{getJavaScriptColumnsIdx}
\alias{getJavaScriptColumnsIdx}
\title{Extract position indices of columns in a data frame and subtract 1}
\usage{
getJavaScriptColumnsIdx(...)
}
\description{
This function is used for correspondance between R and javascript column indices
}
\author{
Michela Pasetto
}
