% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getJsDepMedicalMonitoring}
\alias{getJsDepMedicalMonitoring}
\title{Get Javascript custom scripts required for specific
the medical monitoring functionalities.}
\usage{
getJsDepMedicalMonitoring()
}
\value{
list of \code{\link[htmltools]{htmlDependency}}.
To include this dependency in a report e.g. generated with rmarkdown,
these should be passed to the: \code{extra_dependencies} parameter
of the \code{output_format} specific function, e.g.:
\code{rmarkdown::render(...,	
output_format = rmarkdown::html_document(extra_dependencies = dep))
}
}
\description{
Get Javascript custom scripts required for specific
the medical monitoring functionalities.
}
\author{
Laure Cougnaud
}
