% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_medicalMonitoringReport.R
\name{getMdFromConfig}
\alias{getMdFromConfig}
\title{This function is mainly based from:
\code{bookdown:::insert_chapter_script}}
\usage{
getMdFromConfig(
  configFiles,
  indexPath = "index.Rmd",
  intermediateDir = "./interim"
)
}
\arguments{
\item{configFiles}{Character vector with name or path of the config file(s).}

\item{indexPath}{String with path to the index file,
('index.Rmd' by default).}

\item{intermediateDir}{String with intermediate directory ('interim'
by default), where
markdown files and rds file specifying Js libraries (with \code{knit_meta}) for
each sub report are stored.}

\item{rCmd}{Character vector with R command(s)
to be included in the Rmd document.}

\item{where}{String with position where the R cmd should be included.\cr
Currently, only 'before' is supported.}
}
\value{
No returned value, the Rmd document is updated
such as the command provided in \code{rCmd} 
is included as the first chunk.
Get path of the \code{Markdown} file corresponding
to a specific config file.

The name of the \code{Markdown} file is based on:
\itemize{
\item{for the general \code{config.yml} file: }{
the basename of the specified \code{indexPath}}
\item{for other config file (each sub-report): }{
the name of the config file, after removal of the
'\code{config-}' part.}
}

String with path to the
}
\description{
This function is mainly based from:
\code{bookdown:::insert_chapter_script}
}
\seealso{
Other medical monitoring reporting: 
\code{\link{convertMdToHtml}()},
\code{\link{getMdHeader}()},
\code{\link{getParamsFromConfig}()},
\code{\link{gitbook_medicalMonitoring_report}()},
\code{\link{html_medicalMonitoring_report}()},
\code{\link{knitPrintMedicalMonitoring}()},
\code{\link{render_medicalMonitoringReport}()}
}
\author{
Laure Cougnaud
}
\concept{medical monitoring reporting}
