% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{getMdHeader}
\alias{getMdHeader}
\title{Get Markdown header, for creation in Rmarkdown.}
\usage{
getMdHeader(title, level = 1, settings = NULL)
}
\arguments{
\item{title}{String with header title.}

\item{settings}{List with settings from YAML file, containing:
\itemize{
\item{'rmd\_files': }{character vector with names of Rmd files}
\item{'rmd\_file\_depth': }{integer vector with corresponding file depth}
}}

\item{depth}{Integer of length 1 with header depth/level,
1 by default}
}
\value{
String with Markdown header,
to be included in R within \code{cat}.
}
\description{
The depth is extracted:
\enumerate{
\item{from the \code{settings} if a
specified depth is provided in the 'rmd_file_depth' for the
current knitted file}
\item{\code{depth} parameter otherwise}
}
}
