% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Read metadata file}
\usage{
getMetadata(filePath)
}
\arguments{
\item{filePath}{String of path to file.}
}
\value{
A list of:
\itemize{
\item{\code{pathsInfo}}{ Information extracted from the inputs 
\code{pathSDTMs}, \code{pathMeMoADs} and \code{dateTimeMeMorun}.}
\item{\code{datasetInfo}}{ Information extracted from \code{datasetInfo}.}
}
}
\description{
Read the metadata file from a yaml format.
Currently, this function checks for existance of the following inputs 
provided in yaml:
\itemize{
\item{\code{pathSDTMs}}{ Path to the SDTM data}
\item{\code{pathMeMoADs}}{ Path to the Medical Monitoring Analysis Data sets}
\item{\code{dateTimeMeMorun}}{ Date and time of MeMoADs generation}
\item{\code{datasetInfo}}{ General information about the data sets.}
}
}
