% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_medicalMonitoringReport.R
\name{getParamsFromConfig}
\alias{getParamsFromConfig}
\title{Get parameters from a config file}
\usage{
getParamsFromConfig(configFile, configDir = "./config")
}
\arguments{
\item{configFile}{String with filename of the config
file of interest.}

\item{configDir}{String with directory with config files,
('config' by default).}
}
\value{
List with parameters from the specified \code{configFile}
and the general config file: \code{config.yml}.
}
\description{
Get parameters from a config file
}
\author{
Laure Cougnaud
}
