% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-referenceLines.R
\name{getParFctReferenceLines}
\alias{getParFctReferenceLines}
\title{Get parameter of function used for reference lines}
\usage{
getParFctReferenceLines(type)
}
\arguments{
\item{type}{string with line type, either: 'hline',
'abline' or 'vline'.}
}
\value{
Character vector with parameter names of the functions
}
\description{
Get parameter of function used for reference lines
}
\author{
Laure Cougnaud
}
