% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getPathHyperlink}
\alias{getPathHyperlink}
\title{Get path ('href') property from hyperlink(s).}
\usage{
getPathHyperlink(x)
}
\arguments{
\item{x}{Character vector with hyperlink(s).
If multiple, the hyperlinks should be separated by: ', '.}
}
\value{
Character vector of length \code{x}
containing only the hyperlinks.
}
\description{
Get path ('href') property from hyperlink(s).
}
\author{
Laure Cougnaud
}
