% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlotTableVars.R
\name{getPlotTableVars}
\alias{getPlotTableVars}
\title{Extract variables displayed in the attached table, for
each available plotting function of the medical monitoring package.}
\usage{
getPlotTableVars(plotFunction, plotArgs)
}
\arguments{
\item{plotFunction}{String with name of the plotting function.}

\item{plotArgs}{List with parameters passed to the plotting function.}
}
\value{
Character vector with variable to include in the table,
with extra attribute: 'tableLab'/'tablePars' containing variable labels.
}
\description{
Extract variables displayed in the attached table, for
each available plotting function of the medical monitoring package.
}
\author{
Laure Cougnaud
}
