% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-templates.R
\name{getRdDocFromJSONSch}
\alias{getRdDocFromJSONSch}
\title{Get Roxygen documentation from JSON schema specification
for a specific parameter.}
\usage{
getRdDocFromJSONSch(jsonSch, required = FALSE)
}
\arguments{
\item{jsonSch}{List with JSON Schema
for a specific parameter}

\item{required}{(optional) Logical, if TRUE (FALSE by default)
this parameter is required.}
}
\value{
String with Roxygen documentation.
}
\description{
Get Roxygen documentation from JSON schema specification
for a specific parameter.
}
\author{
Laure Cougnaud
}
