% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-formats.R
\name{gitbook_medicalMonitoring_report}
\alias{gitbook_medicalMonitoring_report}
\title{Medical monitoring format for bookdown report.}
\usage{
gitbook_medicalMonitoring_report(
  split_by = "section+number",
  config = list(sharing = NULL, toc = list(collapse = "section")),
  extra_dependencies = NULL,
  ...
)
}
\arguments{
\item{split_by}{String, how the reports should be split,
(see help of the \link[bookdown]{gitbook} function)}

\item{config}{List with config parameters,
by default: no sharing and collapsed by section.
(see help of the \link[bookdown]{gitbook} function)}

\item{extra_dependencies}{NULL by default}

\item{...}{Extra parameters passed to the
\link[bookdown]{gitbook} function.}
}
\value{
R Markdown output format to pass to \code{\link[bookdown]{render_book}}.
}
\description{
This function is only meant to set sensitive
defaults for gitbook.\cr
\code{\link[bookdown]{gitbook}} can be used instead.
}
\seealso{
Other medical monitoring reporting: 
\code{\link{convertMdToHtml}()},
\code{\link{getMdFromConfig}()},
\code{\link{getMdHeader}()},
\code{\link{getParamsFromConfig}()},
\code{\link{html_medicalMonitoring_report}()},
\code{\link{knitPrintMedicalMonitoring}()},
\code{\link{render_medicalMonitoringReport}()}
}
\author{
Laure Cougnaud
}
\concept{medical monitoring reporting}
