% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-formats.R
\name{html_medicalMonitoring_report}
\alias{html_medicalMonitoring_report}
\title{GLPG-theme medical monitoring format for rmarkdown report.}
\usage{
html_medicalMonitoring_report(
  extra_dependencies = getJsDepMedicalMonitoring(),
  ...
)
}
\arguments{
\item{extra_dependencies}{List of \code{\link[htmltools]{htmlDependency}}
by default extracted from \code{\link{getJsDepMedicalMonitoring}}
(see help of the \link[rmarkdown]{html_document} function)}

\item{...}{Extra parameters passed to the
\link[glpgStyle]{html_report} function.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}.
}
\description{
This format includes also additional javascript libraries
required for visualization/tables of a medical monitoring
report.
}
\seealso{
Other medical monitoring reporting: 
\code{\link{convertMdToHtml}()},
\code{\link{getMdFromConfig}()},
\code{\link{getMdHeader}()},
\code{\link{getParamsFromConfig}()},
\code{\link{gitbook_medicalMonitoring_report}()},
\code{\link{knitPrintMedicalMonitoring}()},
\code{\link{render_medicalMonitoringReport}()}
}
\author{
Laure Cougnaud
}
\concept{medical monitoring reporting}
