% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{includeInButton}
\alias{includeInButton}
\title{Function to create collapse with button in html output}
\usage{
includeInButton(
  input,
  id = paste0("button:", sample.int(n = 1000, size = 1)),
  title = "Click to show or hide",
  color = glpgStyle::glpgColor()["green"],
  borderColor = glpgStyle::glpgColor()["orange"]
)
}
\arguments{
\item{input}{Object to be included within the button.}

\item{id}{String with button ID. If not specified,
a random id, as 'button:x' is used.}

\item{title}{String with button title.}

\item{color}{String witn button color.}

\item{borderColor}{String with button color border.}
}
\value{
\code{\link[htmltools]{HTML}} object
}
\description{
Function to create collapse with button in html output
}
\author{
Kirsten Van Hoorde, Laure Cougnaud
}
