% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{knit_print.medicalMonitoring}
\alias{knit_print.medicalMonitoring}
\title{Print \code{medicalMonitoringTable} object in a knitted document
(e.g. Rmarkdown document).}
\usage{
\method{knit_print}{medicalMonitoring}(x, ...)
}
\arguments{
\item{x}{Object of class \code{medicalMonitoring}}

\item{...}{Extra parameters for \code{\link[knitr]{knit_print}},
not used by default.}
}
\description{
Print \code{medicalMonitoringTable} object in a knitted document
(e.g. Rmarkdown document).
}
\author{
Laure Cougnaud
}
