% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{knitPrintMedicalMonitoring}
\alias{knitPrintMedicalMonitoring}
\title{Include output from medical monitoring, or list 
of such outputs in a Rmarkdown report.}
\usage{
knitPrintMedicalMonitoring(
  list,
  sep = "\\\\.",
  level = 1,
  labelGeneral = "medicalMonitoring"
)
}
\arguments{
\item{list}{List of medical monitoring plots,
potentially nested.
If nested, the list should be named by the different elements,
separated by \code{sep}, e.g. 
\code{list('group1.param1' = .., 'group1.param2' = ...)}.}

\item{sep}{String with separator used to distinguish 
different levels in the labels of the list.
e.g. '\\.' by default.}

\item{level}{Integer with base level for section,
1 by default.}

\item{labelGeneral}{String with general label for the
chunk.}
}
\value{
No returned value, the plots are included in the
report.
}
\description{
Include output from medical monitoring, or list 
of such outputs in a Rmarkdown report.
}
\author{
Laure Cougnaud
}
