% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicalMonitoring-common-args.R
\name{medicalMonitoring-common-args}
\alias{medicalMonitoring-common-args}
\title{Common arguments for the functions of the medicalMonitoring package}
\arguments{
\item{data}{Data.frame with data.}

\item{gg}{\code{\link[ggplot2]{ggplot}} object.}

\item{xVar}{String with column of \code{data} containing x-variable.}

\item{yVar}{String with column of \code{data} containing y-variable.}

\item{xLim, yLim}{Numeric vector of length 2 with limits for the x/y axes.}

\item{idVar}{String with variable containing subject ID.}

\item{idLab}{String with label for \code{idVar}.}

\item{facetPars}{List with facetting parameters, passed to the facetting function.}

\item{lineVars}{List with parameters for the reference lines.}

\item{hoverVar}{Character vector with variables to be displayed in the hover,
by default \code{xVar}, \code{yVar} and any aesthetic variables.}

\item{hoverLab}{Named character vector with labels for \code{hoverVar}.}

\item{pathVar}{String with variable of \code{data} containing path
to a subject-specific report (e.g. patient profiles).
The report should be unique by element of \code{idVar}.
This report will be:p
\itemize{
\item{downloaded if the user clicks on the 'p' (a.k.a 'profile') key
when hovering on a point of the plot}
\item{opened in a browser via hyperlink in the \code{idVar} of the table 
column (if specified via \code{tableVars})}
}}

\item{table}{Logical, if TRUE (FALSE by default)
returns also a \code{datatable} containing the plot data.}

\item{refLinePars}{(optional) Nested list, with parameters for each reference line(s).
Each sublist (a.k.a reference line) contains:
\itemize{
\item{aesthetic value(s) or variable(s) for the lines
(in this case column names of \code{data}) for reference lines.
The line position is controlled by the aesthetics supported in
\code{\link[ggplot2]{geom_vline}}, \code{\link[ggplot2]{geom_hline}} 
and \code{\link[ggplot2]{geom_abline}}.
}
\item{'label': }{(optional) Logical specifying if the line
should be annotated (\code{FALSE} to not annotate the line)
or string with annotation label. By default, the value
of the position of the horizontal/vertical line or the equation
of the diagonal line is displayed.
}
}}

\item{labelVars}{Named character vector containing variable labels,
used by default for all labels in the plot.}

\item{id}{String with general id for the plot:
\itemize{
\item{'id' is used as \code{group} for the \code{\link[crosstalk]{SharedData}}}
\item{'button:[id]' is used as button ID if \code{table} is TRUE}
}
If not specified, a random id, as 'plotMonitoring[X]' is used.}

\item{title}{String with title for the plot.}

\item{titleExtra}{String with extra title for the plot (appended after \code{title}).}

\item{width, height}{Width/height of the plot in pixels.}
}
\description{
Common arguments for the functions of the medicalMonitoring package
}
