% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeHtml.R
\name{parseScrapeTable}
\alias{parseScrapeTable}
\title{Parse JSON objects}
\usage{
parseScrapeTable(jsonObj)
}
\arguments{
\item{jsonObj}{An object from \code{fromJSON}.
The object must contain
\itemize{
   \item{x}{
       \item{data}{The data}
       \item{container}{The container to extract the column names
to assign to the data}
  }
}}
}
\value{
A matrix object, as a table of the data
}
\description{
Parsing JSON objects from \code{jsonlite::fromJSON} function.
See functions \code{readHtml}.
}
\author{
Michela Pasetto
}
