% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{printMetadataInReport}
\alias{printMetadataInReport}
\title{Print metadata file in the medical monitoring report}
\usage{
printMetadataInReport(metadataInfo)
}
\arguments{
\item{metadataInfo}{List of two elements named \code{pathsInfo} and 
\code{datasetInfo}.}
}
\value{
Nothing. The tables are ready to be printed in Rmd.
}
\description{
This function receives the metadata information from \code{\link{getMetadata}} and 
prints them in a format for an Rmd report.
In general, any list could be called as long as it is composed by two elements:
\itemize{
\item{\code{pathsInfo}}{ an R object.}
\item{\code{datasetInfo}}{ a data.frame or a matrix.}
}
The first (\code{pathsInfo}) is printed as \code{\link{knitr::kable}} object 
and the second (\code{datasetInfo}) is printed as hide/show html button with 
the function \code{\link{collapseHtmlContent}}.
}
