% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeHtml.R
\name{readHtml}
\alias{readHtml}
\title{Read html}
\usage{
readHtml(pathHtml)
}
\arguments{
\item{pathHtml}{String pointing to the path
where the html is stored}
}
\value{
A JSON object
}
\description{
Read the html file specified.
}
\author{
Michela Pasetto
}
