% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{renamePathDateInfoMetadata}
\alias{renamePathDateInfoMetadata}
\title{Rename variable names of metadata info}
\usage{
renamePathDateInfoMetadata(pathsInfo)
}
\arguments{
\item{pathsInfo}{A matrix, see output from \code{\link{getMetadata}}.}
}
\value{
A matrix, same as input \code{pathsInfo} with renamed variable names.
}
\description{
Rename variable names referring to the paths and the date.
}
