% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_medicalMonitoringReport.R
\name{render_medicalMonitoringReport}
\alias{render_medicalMonitoringReport}
\title{Render a medical monitoring report.}
\usage{
render_medicalMonitoringReport(
  indexPath = "index.Rmd",
  outputDir = "./MOMP",
  configDir = "./config",
  extraDirs = c("figures", "tables"),
  ...
)
}
\arguments{
\item{indexPath}{String with path to the index file,
('index.Rmd' by default).}

\item{outputDir}{String with output directory,
('MOMP', for: 'Medical Oversight and monitoring Plan' by default).}

\item{configDir}{String with directory with config files,
('config' by default).
It should contain a general 'config.yml' file and dedicated
'config-[X].yml' for each chapter.
The order of each chapter is specified in the 'config' slot in the general 
general 'config.yml'.}

\item{extraDirs}{Character vector with extra directories required by
the report, directory with external images  .
By default, the directories: 'figures', 'tables' and 
directory included in the 'patientProfilePath' parameter of the
general config file are considered.}

\item{...}{Extra parameters for the link[rmarkdown]{render} function.}
}
\value{
String with path to the medical monitoring report.
}
\description{
This function is based on the \link[bookdown]{render_book}
function, enabling specification of chapter-specific input parameters,
specified in YAML configuration files.
}
\author{
Laure Cougnaud
}
