% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPatientProfileTemplateReport.R
\name{runPatientProfileTemplateReport}
\alias{runPatientProfileTemplateReport}
\title{Create patient profile report, based on template.}
\usage{
runPatientProfileTemplateReport(
  dataPath,
  overwrite = FALSE,
  outputFile = "subjectProfile.pdf",
  study = "custom",
  batch = "X",
  author = NULL,
  subsetDataset = NULL,
  subsetVar = NULL,
  subsetValue = NULL,
  subsetSample = NULL
)
}
\arguments{
\item{dataPath}{String with path to data.}

\item{overwrite}{Logical, if TRUE (FALSE by default)
the patient profile template is overwritten in the 
output directory.}

\item{outputFile}{(optional) String with name of output file,
'subjectProfile.pdf' by default.}

\item{study}{(optional) String with study name.}

\item{batch}{(optional) String with batch name.}

\item{author}{(optional) String with author name.}
}
\value{
Character vector with output path.
The patient profile(s) and the Rmd template
are created at \code{outputFile}.
}
\description{
Currently template is only available for SDTM data format.
}
\author{
Laure Cougnaud
}
