% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeHtml.R
\name{saveScrapeTable}
\alias{saveScrapeTable}
\title{Save scraped table}
\usage{
saveScrapeTable(table, pathToDir, fileName)
}
\arguments{
\item{table}{A matrix object to be saved}

\item{pathToDir}{String with the path to directory}

\item{fileName}{String with name of the file to be saved}
}
\value{
The table is saved as both txt and csv file at the
directory indicated with \code{pathToDir} in a folder called 
"scrapedTables".
}
\description{
Save a table to a path.
}
\author{
Michela Pasetto
}
