% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeHtml.R
\name{scrapeHtml}
\alias{scrapeHtml}
\title{Scrape html}
\usage{
scrapeHtml(pathHtml)
}
\arguments{
\item{pathHtml}{String pointing to the path
where the html is stored}
}
\value{
The data
}
\description{
Scrape html file(s) and retrieve the data
}
\examples{

#' \dontrun{
## Demographics
#demographicsHtml <- "~/Documents/Projects/Galapagos/medicalMonitoring/demoReleases/20200703/MOMP/3-3-demographics.html"
#
### Listings
#aeListingHtml <- "~/Documents/Projects/Galapagos/medicalMonitoring/demoReleases/20200703/MOMP/5-6-adverse-events-listing.html"
#
#tableHtml <- scrapeHtml(pathHtml = aeListingHtml)

#' }

}
\author{
Michela Pasetto
}
