% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-facets.R
\name{setFacetLayoutWrap}
\alias{setFacetLayoutWrap}
\title{Set facetting layout for 'wrap' facetting.}
\usage{
setFacetLayoutWrap(data, facetPars = list())
}
\value{
Updated \code{facetPars}.
}
\description{
By default, the number of columns is 2.
}
\author{
Laure Cougnaud
}
