% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticScatterplotMonitoring.R
\name{setPaletteStaticScatterplotMonitoring}
\alias{setPaletteStaticScatterplotMonitoring}
\title{Get standard palette for the \code{staticScatterplotMonitoring}
function.}
\usage{
setPaletteStaticScatterplotMonitoring(data, var, aes, scalePars, geomAes, ...)
}
\arguments{
\item{data}{Data.frame with data for the plot.}

\item{var}{Character vector with variable(s) to consider.
If multiple, currently only the first one is considered.}

\item{aes}{String with aesthetic, either:
'color', 'shape' or 'linetype'.}

\item{scalePars}{List with parameters to customize
scales. Each sublist should contains a set of parameters
passed to the \code{\link[ggplot2]{scale_discrete_manual}}
function.}

\item{geomAes}{List with aesthetic for each geom.}

\item{...}{Any extra parameters than \code{x} and \code{n}
for the default palette fcts.}
}
\value{
List with: \code{scalePars} and \code{geomAes},
each of those potentially updated with default palette(s).
}
\description{
Get standard palette for the \code{staticScatterplotMonitoring}
function.
}
\author{
Laure Cougnaud
}
