% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticPlotMonitoring.R
\name{staticScatterplotMonitoring}
\alias{staticScatterplotMonitoring}
\title{Scatterplot of variables of interest for medical monitoring.}
\usage{
staticScatterplotMonitoring(
  data,
  xVar,
  yVar,
  xLab = getLabelVar(xVar, labelVars = labelVars),
  yLab = getLabelVar(yVar, labelVars = labelVars),
  aesPointVar = list(),
  aesLineVar = list(),
  aesLab,
  xTrans = "identity",
  yTrans = "identity",
  xPars = list(),
  yPars = list(),
  yLim = NULL,
  xLim = NULL,
  titleExtra = NULL,
  title = paste(paste(yLab, "vs", xLab, titleExtra), collapse = "<br>"),
  facetPars = list(),
  facetType = c("wrap", "grid"),
  themePars = list(legend.position = "bottom"),
  refLinePars = NULL,
  labelVars = NULL,
  hoverVars = NULL,
  geomType = c("point", "col")
)
}
\arguments{
\item{data}{Data.frame with input data.}

\item{xVar}{String with column of \code{data} containing x-variable.}

\item{yVar}{String with column of \code{data} containing y-variable.}

\item{xLab}{String with label for \code{xVar}.}

\item{yLab}{String with label for \code{xVar}.}

\item{aesPointVar}{List with specification of aesthetic variable(s),
for the point, passed to the \code{mapping} parameter of \code{\link[ggplot2]{geom_point}},
e.g. \code{list(color = "TRTP")}.}

\item{aesLineVar}{List with specification of aesthetic variable(s),
for the line, passed to the \code{mapping} parameter of \code{\link[ggplot2]{geom_point}},
e.g. \code{list(group = "USUBJID")}.}

\item{aesLab}{Named character vector with labels for each aesthetic variable.}

\item{xTrans, yTrans}{Transformation for the x/y- variables,
passed to the \code{trans} parameter of \code{\link[ggplot2]{scale_x_continuous}}/
\code{\link[ggplot2]{scale_y_continuous}}.}

\item{xPars, yPars}{List with extra parameters for x/y axis, passed to the
\code{\link[ggplot2]{scale_x_continuous}}/
\code{\link[ggplot2]{scale_y_continuous}} functions,
besides \code{trans} and \code{limits}.}

\item{yLim}{Numeric vector of length 2 with limits for the x/y axes.}

\item{xLim}{Numeric vector of length 2 with limits for the x/y axes.}

\item{titleExtra}{String with extra title for the plot (appended after \code{title}).}

\item{title}{String with title for the plot.}

\item{facetPars}{List with facetting parameters, passed to the facetting function.}

\item{facetType}{String with facetting type, either:
\itemize{
\item{'wrap': }{\code{\link[ggplot2]{facet_wrap}}}
\item{'grid': }{\code{\link[ggplot2]{facet_grid}}}
}}

\item{themePars}{List with general theme parameters 
(see \code{\link[ggplot2]{theme}}).}

\item{refLinePars}{(optional) Nested list, with parameters for each reference line(s).
Each sublist (a.k.a reference line) contains:
\itemize{
\item{aesthetic value(s) or variable(s) for the lines
(in this case column names of \code{data}) for reference lines.
The line position is controlled by the aesthetics supported in
\code{\link[ggplot2]{geom_vline}}, \code{\link[ggplot2]{geom_hline}} 
and \code{\link[ggplot2]{geom_abline}}.
}
\item{'label': }{(optional) Logical specifying if the line
should be annotated (\code{FALSE} to not annotate the line)
or string with annotation label. By default, the value
of the position of the horizontal/vertical line or the equation
of the diagonal line is displayed.
}
}}

\item{labelVars}{Named character vector containing variable labels,
used by default for all labels in the plot.}

\item{hoverVars}{Character vector with variables to be displayed in the hover,
by default \code{xVar}, \code{yVar} and any aesthetic variables.}

\item{geomType}{String with type of the geom used, either:
'point' or 'col'}
}
\value{
\code{\link[ggplot2]{ggplot}} object
}
\description{
Scatterplot of variables of interest for medical monitoring.
}
\author{
Laure Cougnaud
}
