% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMonitoring.R
\name{tableMonitoring}
\alias{tableMonitoring}
\title{Create a 'monitoring table', associated to a plot.}
\usage{
tableMonitoring(
  data,
  idVar = "USUBJID",
  idLab = getLabelVar(idVar, labelVars = labelVars),
  pathVar = NULL,
  pathLab = getLabelVar(pathVar, labelVars = labelVars),
  pathExpand = FALSE,
  tableVars,
  tableLab = getLabelVar(tableVars, labelVars = labelVars),
  tableButton = TRUE,
  tablePars = list(),
  id = paste0("plotMonitoring", sample.int(n = 1000, size = 1)),
  labelVars = NULL
)
}
\arguments{
\item{data}{Data.frame with data.}

\item{idVar}{String with variable containing subject ID.}

\item{idLab}{String with label for \code{idVar}.}

\item{pathVar}{String with variable of \code{data} containing:
\itemize{
\item{visualization of individual profiles 
(\code{\link{scatterplotMonitoring}}) :}{
path to the report. The report should be unique 
for each element of \code{idVar}.
}
\item{visualization of summary statistics 
(\code{\link{sunburstMonitoring}}), \code{\link{treemapMonitoring}},
\code{\link{barplotMonitoring}}) :}{path or hyperlinks 
with path to the report. If multiple, they should be separated by: ', '.
}
}
The report(s) will be:
\itemize{
\item{downloaded or opened in the browser if the user clicks on the 'p' (a.k.a 'profile') key
when hovering on a point of the plot}
\item{downloaded or opened in the browser via hyperlink in the table}
}}

\item{pathLab}{String with label for \code{pathVar},
used for the label of the collapsible row
(only displayed if \code{pathExpand} is TRUE).}

\item{pathExpand}{Logical, should the variable in \code{pathExpand}
be included in a collapsible row or as hyperlink in the table?
Should be TRUE for if multiple paths are included for each \code{idVar},
FALSE otherwise (by default).}

\item{tableVars, tableLab}{Character vector with variables to be included 
in the table; and associated labels.}

\item{tableButton}{Logical, if TRUE (by default)
the table is included within an HTML button.}

\item{tablePars}{List with parameters passed to the
\code{\link[glpgUtilityFct]{toDTGLPG}} function.}

\item{id}{String with general id for the plot:
\itemize{
\item{'id' is used as \code{group} for the \code{\link[crosstalk]{SharedData}}}
\item{'button:[id]' is used as button ID if \code{table} is TRUE}
}
If not specified, a random id, as 'plotMonitoring[X]' is used.}

\item{labelVars}{Named character vector containing variable labels,
used by default for all labels in the plot.}
}
\value{
\code{\link[DT]{datatable}}
}
\description{
Interactive table is created, with the possibility to have
clickeable link to patient-specific report, and included
within a button.
}
\author{
Laure Cougnaud
}
