% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_svg_plotly.R
\name{write_svg_plotly}
\alias{write_svg_plotly}
\title{Export a \code{\link[plotly]{plotly}} object to a svg file.}
\usage{
write_svg_plotly(plot, file, title = "")
}
\arguments{
\item{plot}{\code{\link[plotly]{plotly}}}

\item{file}{String with path to file where the figure should be exported.}

\item{title}{(optional) String with figure title,
this will be added to the current plot title (if present), 
e.g. to label a specific test.}
}
\value{
No returned value, the \code{plot} is saved to the
\code{file}.
}
\description{
This function is not intended to be used directly, but 
within the package unit tests,
and be passed to the \code{writer} parameter of the
\code{\link[vdiffr]{expect_doppelganger}} function.
}
\references{
Similar functionality is available in the plotly package 
in the 'tests': 'write_plotly_svg', but not exported (github issue #1703).
}
\author{
Laure Cougnaud
}
