% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipMedicalMonitoring.R
\name{zipMedicalMonitoring}
\alias{zipMedicalMonitoring}
\title{Zip the medical monitoring report}
\usage{
zipMedicalMonitoring(
  reportDir = "./report",
  newDir = "./report_dependencies",
  redirectPage = "report.html",
  zipFolder = "report.zip"
)
}
\arguments{
\item{reportDir}{String for the path to the directory where
the medical monitoring reports are stored}

\item{newDir}{String for the path where the files from
\code{reportDir} should be copied to.}

\item{redirectPage}{String with the path of the html file that redirects
to the "1-introduction.html" page of the report.}

\item{zipFolder}{String with the path to the zipped folder.}
}
\value{
The zip folder is created in the specified location.
}
\description{
Create a zip folder of medical monitoring reports with a redirect page.
The medical monitoring report out of the 
\code{\link{render_medicalMonitoring}} is copied into a new folder.
A redirect html page is created to enable the user to navigate the report
without needing to look into the new directory.
}
