% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alertData.R
\name{alertData}
\alias{alertData}
\title{Create alert data}
\usage{
alertData(
  data = "adsl",
  dataPath = ".",
  alerts,
  subjectVar = "USUBJID",
  verbose = FALSE,
  labelVars = NULL,
  labelData = "data"
)
}
\arguments{
\item{data}{Data.frame with data to create the subject level alerts.
By default 'adsl'.}

\item{dataPath}{String with path to the data.}

\item{subjectVar}{String with subject ID variable, 'USUBJID' by default.}

\item{verbose}{Logical, if TRUE (FALSE by default) progress messages are printed
in the current console.}

\item{labelVars}{Named character vector containing variable labels of \code{data}.
This will be updated with the labels of the extra annotation variables
(in \code{attr(output, 'labelVars')}).}

\item{labelData}{(optional) String with label for input \code{data},
that will be included in progress messages.}

\item{alert}{A list of custom alerts as:
\itemize{
        \item{(optional) annotation dataset, either:
        \item{'dataset': }{
            String with name of the annotation dataset,
         e.g. 'ex' to import data from the file: '[dataset].sas7bdat'in \code{dataPath}
            }
        \item{'data': }{
            Data.frame with annotation dataset
            }
        }
        \item{'vars': }{
            Character vector with variables of interest from annotation dataset
            }
        \item{'varFct': }{(optional) Function of \code{data} or string containing
            manipulations from column names of \code{data} used to 
        create a new variable specified in \code{vars}.}
        \item{'filters': }{(optional) Filters for the annotation dataset, 
            see \code{filters} parameter of \code{\link{filterData}}}
        \item{'varLabel': }{(optional) label for new variable in case \code{varFct} is specified.}
        \item{'varsBy': }{(optional) Character vector with variables used to merge input data and
            the annotation dataset. If not specified, \code{subjectVar} is used if
            an external annotation dataset, or the datasets are merged by rows otherwise.
}
}}
}
\description{
Create alert data
}
\author{
Michela Pasetto
}
