% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alertData.R
\name{alertData}
\alias{alertData}
\title{Create alert data}
\usage{
alertData(
  data,
  dataPath = ".",
  alerts,
  subjectVar = "USUBJID",
  verbose = FALSE,
  labelVars = NULL,
  labelData = "data"
)
}
\arguments{
\item{data}{Data.frame with data to create the alerts.}

\item{dataPath}{String with path to the data.}

\item{alerts}{A list of custom alerts as:
\itemize{
        \item{'vars': }{
            Character vector with (new) variables of alerts
            }
        \item{'varFct': }{Function of \code{data} based on column names of \code{data} used to 
        create the alert specified in \code{vars}. This function should provide a logical evaluation.}
        \item{'filters': }{(optional) Filters for the \code{data}, 
            see \code{filters} parameter of \code{\link{filterData}}}
        \item{'varLabel': }{(optional) label for new variable. Otherwise the name from \code{varFct} is specified.}
}}

\item{subjectVar}{String with subject ID variable, 'USUBJID' by default.}

\item{verbose}{Logical, if TRUE (FALSE by default) progress messages are printed
in the current console.
For the visualizations, progress messages during download
of subject-specific report are displayed in the browser console.}

\item{labelVars}{Named character vector containing variable labels of \code{data}.
This will be updated with the labels of the extra annotation variables
(in \code{attr(output, 'labelVars')}).}

\item{labelData}{(optional) String with label for input \code{data},
that will be included in progress messages.}
}
\description{
Create alert data
}
\author{
Michela Pasetto
}
