% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateData.R
\name{annotateData}
\alias{annotateData}
\title{Annotate a SDTM dataset.}
\usage{
annotateData(
  data,
  dataPath = ".",
  annotations,
  subjectVar = "USUBJID",
  verbose = FALSE,
  labelVars = NULL,
  labelData = "data"
)
}
\arguments{
\item{data}{Data.frame with input data to annotate.}

\item{dataPath}{String with path to the data.}

\item{annotations}{Annotations (or list of those) either as a:
\itemize{
\item{string with standard annotation type, among:
\itemize{
\item{demographics: }{standard variables from the demographics data (DM or ADSL) are extracted}
\item{exposed_subjects: }{a logical variable: \code{EXFL} is added to \code{data},
identifying exposed subjects, i.e. subjects included in the exposure dataset (EX/ADEX)
dataset and with non empty and non missing start date ('EXSTDTC', 'STDY' or 'ASTDY')}
\item{functional_groups_lab: }{a character variable: 'LBFCTGRP' is added to \code{data}
based on standard naming of the parameter code ('PARAMCD' or 'LBTESTCD' variable)}
}
}
\item{list of custom annotation, with:
\itemize{
\item{(optional) annotation dataset, either:
\item{'dataset': }{String with name of the annotation dataset,
 e.g. 'ex' to import data from the file: '[dataset].sas7bdat'in \code{dataPath}}
\item{'data': }{Data.frame with annotation dataset}
The input \code{data} is used if 'data' and 'dataset' are not specified.}
\item{'vars': }{Character vector with variables of interest from annotation dataset.
If not specified, all variables of the dataset are considered.}
\item{'varFct': }{(optional) Function of \code{data} or string containing
manipulations from column names of \code{data} used to 
create a new variable specified in \code{vars}.}
\item{'filters': }{(optional) Filters for the annotation dataset, 
see \code{filters} parameter of \code{\link{filterData}}}
\item{'varsBy': }{(optional) Character vector with variables used to merge input data and
the annotation dataset. If not specified, \code{subjectVar} is used if
an external annotation dataset, or the datasets are merged by rows otherwise.
}
}
}
}}

\item{subjectVar}{String with subject ID variable, 'USUBJID' by default.}

\item{verbose}{Logical, if TRUE (FALSE by default) progress messages are printed
in the current console.}

\item{labelVars}{Named character vector containing variable labels of \code{data}.
This will be updated with the labels of the extra annotation variables
(in \code{attr(output, 'labelVars')}).}

\item{labelData}{(optional) String with label for input \code{data},
that will be included in progress messages.}
}
\value{
Annotated \code{data}.
If \code{labelVars} is specified, the output contains an 
extra attribute: 'labelVars'
containing updated \code{labelVars} (accessible via: in \code{attr(output, 'labelVars')}).
}
\description{
Standard annotation variables are available via the 
parameter \code{annotType}. Custom dataset/variables of interest
are specified via the \code{annotDataset}/\code{annotVar} parameters.
}
\examples{
library(glpgUtilityFct)

data(SDTMDataPelican)

dataLB <- SDTMDataPelican$LB
dataDM <- SDTMDataPelican$DM
dataEX <- SDTMDataPelican$EX
dataAE <- SDTMDataPelican$AE

data(labelVarsSDTMPelican)
labelVars <- labelVarsSDTMPelican

# standard annotations:
# path to dataset should be specified via: 'pathData'
\dontrun{
annotateData(dataLB, annotations = "demographics", pathData = ...)
}

# add all variables in annotation data (if not already available)
head(annotateData(dataLB, annotations = list(data = dataDM)), 1)

# only variables of interest
head(annotateData(dataLB, annotations = list(data = dataDM, vars = c("ARMCD", "COUNTRY"))), 1)

# filter annotation dataset
dataAnnotated <- annotateData(dataLB, 
	annotations = list(
		data = dataDM, 
		vars = c("ARMCD", "COUNTRY", "SEX"), 
		filters = list(var = "SEX", value = "F"))
	)
head(subset(dataAnnotated, SEX == "F"), 1)
head(subset(dataAnnotated, is.na(SEX)), 1)

# worst-case scenario: add a new variable based on filtering condition
dataAE$AESEV <- factor(dataAE$AESEV, levels = c('MILD', "MODERATE"))
dataAEWC <- annotateData(
	data = dataAE,
	annotations = list(
		vars = "WORSTINT", 
		# create new variable: 'WORSTINT' 
		# with TRUE if maximum toxicity grade per subject/test 
		# (if multiple, they are all retained)
		filters = list(
			var = "AESEV", 
			valueFct = function(x) x[which.max(as.numeric(x))],
			varsBy = c("USUBJID", "AEDECOD"),
			keepNA = FALSE,
			varNew = "WORSTINT", 
			labelNew = "worst-case"
		)
	),
	labelVars = labelVars,
	verbose = TRUE
)
attr(dataAEWC, "labelVars")["WORSTINT"]

# add a new variable based on a combination of variables:
dataLB <- annotateData(dataLB, 
	annotations = list(vars = "ULN_ratio", varFct = "LBSTRESN / LBSTNRHI")
)

# add a new variable based on extraction of a existing variable
# Note: slash should be doubled when the function is specified as text
dataLB <- annotateData(dataLB, 
	annotations = list(vars = "PERIOD", varFct = "sub('Day .* - (.+)', '\\\\\\\\1', VISIT)")
)
print(unique(dataLB[, c("VISIT", "PERIOD")]))

# multiple annotations:
dataAnnotated <- annotateData(dataLB, 
	annotations = list(
		list(data = dataDM, vars = c("ARMCD", "COUNTRY", "SEX")),
		list(data = dataEX)
	)
)
head(dataAnnotated, 1)
head(subset(dataAnnotated, !is.na(EXSEQ)), 1)
}
