% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{collapseHtmlContent}
\alias{collapseHtmlContent}
\title{Function to create collapsible HTML content}
\usage{
collapseHtmlContent(input, title = "Click to show or hide")
}
\arguments{
\item{input}{Object to be collapse, e.g.
datatable.}

\item{title}{String with button title.}
}
\value{
\code{\link[htmltools]{tag}} object
}
\description{
Please note that the button is of class:
'hideshow', defined in the 'input.hideshow.js' js file
included in the package.
}
\author{
Laure Cougnaud
}
