% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-formatData.R
\name{formatDataForPlotMonitoring}
\alias{formatDataForPlotMonitoring}
\title{Format data for interactive plot for medical monitoring.}
\usage{
formatDataForPlotMonitoring(
  data,
  hoverVars = NULL,
  hoverLab = getLabelVar(hoverVars, labelVars = labelVars),
  hoverByVar = NULL,
  keyVar = NULL,
  id = paste0("plotMonitoring", sample.int(n = 1000, size = 1)),
  labelVars = NULL
)
}
\arguments{
\item{hoverByVar}{Character vector with variables identifying
unique elements in the plot, usually x, y, facet variables.
These variables are used to identify records with the same position
in the plot, their information are combined in the hover.}

\item{keyVar}{Character vector with key variables, identifying unique
group for which the link between the table and the plot should be done.}
}
\value{
Updated \code{\link[crosstalk]{SharedData}} with:
\itemize{
\item{extra column: 'hover' with combined info from \code{hoverVars}}
}
}
\description{
Format data for interactive plot for medical monitoring.
}
\author{
Laure Cougnaud
}
