% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-metadata.R
\name{formatPathDateInfoMetadata}
\alias{formatPathDateInfoMetadata}
\title{Format the info on paths from metadata}
\usage{
formatPathDateInfoMetadata(summaryInfo, namesInfo)
}
\arguments{
\item{summaryInfo}{matrix, see output from \code{\link{getMetadata}}.}

\item{namesInfo}{Named vector to rename the final output.}
}
\value{
A kable object, to be printed.
}
\description{
Format the info on paths from metadata
}
