% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-formatPlotlyMonitoring.R
\name{formatPlotlyMonitoring}
\alias{formatPlotlyMonitoring}
\title{Format interactive plot,
with possibility to download patient profiles
on a click event.}
\usage{
formatPlotlyMonitoring(
  pl,
  data,
  idVar = "USUBJID",
  pathVar = NULL,
  idFromDataPlot = TRUE,
  idVarPlot = "key",
  highlightOn = "plotly_click",
  highlightOff = "plotly_doubleclick",
  id = paste0("plotMonitoring", sample.int(n = 1000, size = 1))
)
}
\arguments{
\item{pl}{\code{\link[plotly]{ggplotly}} object.}

\item{idVar}{String with variable of \code{data}
containing plot element.}

\item{idFromDataPlot}{Logical, if TRUE (by default) \code{idVarPlot}
is extracted from the data of the plot object, otherwise
directly from the plot object.}

\item{idVarPlot}{String with variable in the \code{\link[plotly]{plotly}}
output containing IDs.}

\item{highlightOn}{String with event to turn on the selection
(\code{on} parameter of \code{\link[plotly]{highlight}}),
'plotly_click' by default.}

\item{highlightOff}{String with event to turn off the selection
(\code{off} parameter of \code{\link[plotly]{highlight}}),
'plotly_doubleclick' by default.}
}
\value{
Updated \code{\link[plotly]{plotly}} object.
}
\description{
Format interactive plot,
with possibility to download patient profiles
on a click event.
}
\author{
Laure Cougnaud
}
