% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-formatToHierarchicalData.R
\name{formatToHierarchicalData}
\alias{formatToHierarchicalData}
\title{Format data to a hierarchical data, in the format
as required by the plotly sunburst and treemap.}
\usage{
formatToHierarchicalData(data, vars)
}
\arguments{
\item{data}{Data.frame with data.}

\item{vars}{Character vector with variable(s) of interest.}
}
\value{
Updated data.frame with \code{vars} in 
hierarchical format, with extra attributes (in 'metadat'):
\itemize{
\item{'varsPlotly': }{Named character vector with
new variable names (\code{[var].hier}) names
with original variable names
}
\item{'varLabel': }{String with name
of new column containing text used for
label in plotly.}
}
}
\description{
Note that new variables are created for each variable of interest
(the variables are not overwritten) to avoid
issues with cases where the value in the child 
and parent variables are the same.
}
\author{
Laure Cougnaud
}
