% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-facets.R
\name{getFacetVars}
\alias{getFacetVars}
\title{Get facetting variables from facet parameters.}
\usage{
getFacetVars(facetPars = list())
}
\arguments{
\item{facetPars}{List with facetting parameters, 
passed to the facetting function. Variables should be specified
as character or formula.
For 'wrap' facetting (\code{facetType} is 'wrap'), if the layout
is not specified via \code{nrow}/\code{ncol}, 2 columns
are used by default.}
}
\value{
Character vector with facetting variable
}
\description{
Get facetting variables from facet parameters.
}
\author{
Laure Cougnaud
}
