% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-utility.R
\name{getMdHeader}
\alias{getMdHeader}
\title{Get Markdown header, for creation in Rmarkdown.}
\usage{
getMdHeader(title, level = 1, settings = NULL)
}
\arguments{
\item{title}{String with header title.}

\item{level}{Integer of length 1 with header depth/level,
1 by default}

\item{settings}{List with settings from YAML file, containing:
\itemize{
\item{'rmd\_files': }{character vector with names of Rmd files}
\item{'rmd\_file\_depth': }{integer vector with corresponding file depth}
}}
}
\value{
String with Markdown header,
to be included in R within \code{cat}.
}
\description{
The depth is extracted:
\enumerate{
\item{from the \code{settings} if a
specified depth is provided in the 'rmd_file_depth' for the
current knitted file}
\item{\code{level} parameter otherwise}
}
}
\seealso{
Other medical monitoring reporting: 
\code{\link{convertMdToHtml}()},
\code{\link{getMdFromConfig}()},
\code{\link{getParamsFromConfig}()},
\code{\link{gitbook_medicalMonitoring_report}()},
\code{\link{html_medicalMonitoring_report}()},
\code{\link{knitPrintMedicalMonitoring}()},
\code{\link{render_medicalMonitoringReport}()}
}
\concept{medical monitoring reporting}
