% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_medicalMonitoringReport.R
\name{getParamsFromConfig}
\alias{getParamsFromConfig}
\title{Get parameters from a config file}
\usage{
getParamsFromConfig(configFile, configDir = "./config")
}
\arguments{
\item{configFile}{String with filename of the config
file of interest in YAML format.}

\item{configDir}{String with directory with config files,
('config' by default).
It should contain a general 'config.yml' file and dedicated
'config-[X].yml' for each chapter.
The order of each chapter is specified in the 'config' slot in the general 
general 'config.yml'.}
}
\value{
List with parameters from the specified \code{configFile}
and the general config file: \code{config.yml}.\cr
There are two specific handlers:
\itemize{
\item{parameters tagged with '[param] !r [value]'
are evaluated in R, and their evaluated value is returned}
\item{parameters tagged with '[param] !r-lazy [value]'
are imported as character, and need to be further process
with \code{\link{forceParams}} inside the report.}
}
Parameters with YAML type 'r-lazy' are imported as character,
with this additional class.
}
\description{
Get parameters from a config file
}
\seealso{
\link{forceParams}

Other medical monitoring reporting: 
\code{\link{convertMdToHtml}()},
\code{\link{getMdFromConfig}()},
\code{\link{getMdHeader}()},
\code{\link{gitbook_medicalMonitoring_report}()},
\code{\link{html_medicalMonitoring_report}()},
\code{\link{knitPrintMedicalMonitoring}()},
\code{\link{render_medicalMonitoringReport}()}
}
\author{
Laure Cougnaud
}
\concept{medical monitoring reporting}
