% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-formats.R
\name{gitbook_medicalMonitoring_report}
\alias{gitbook_medicalMonitoring_report}
\title{GLPG-theme medical monitoring format for bookdown report.}
\usage{
gitbook_medicalMonitoring_report(
  split_by = "section+number",
  config = list(sharing = NULL, toc = list(collapse = "section")),
  extra_dependencies = getJsDepMedicalMonitoring(),
  ...
)
}
\arguments{
\item{split_by}{String, how the reports should be split,
(see help of the \link[bookdown]{gitbook} function)}

\item{config}{List with config parameters,
by default: no sharing and collapsed by section.
(see help of the \link[bookdown]{gitbook} function)}

\item{extra_dependencies}{List of \code{\link[htmltools]{htmlDependency}}
by default extracted from \code{\link{getJsDepMedicalMonitoring}}.
(see help of the \link[rmarkdown]{html_document} function)}

\item{...}{Extra parameters passed to the
\link[glpgStyle]{gitbook_report} function.}
}
\value{
R Markdown output format to pass to \code{\link[bookdown]{render_book}}.
}
\description{
This format includes also additional javascript libraries
required for visualization/tables of a medical monitoring
report.
}
\author{
Laure Cougnaud
}
