% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-formats.R
\name{html_medicalMonitoring_report}
\alias{html_medicalMonitoring_report}
\title{Medical monitoring format for rmarkdown report.}
\usage{
html_medicalMonitoring_report(extra_dependencies = NULL, ...)
}
\arguments{
\item{extra_dependencies}{NULL by default.}

\item{...}{Extra parameters passed to the
\link[rmarkdown]{html_document} function.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}.
}
\description{
This function only kept for back-compatibility, 
\code{\link[rmarkdown]{html_document}}
can be used instead.
}
\seealso{
Other medical monitoring reporting: 
\code{\link{convertMdToHtml}()},
\code{\link{getMdFromConfig}()},
\code{\link{getMdHeader}()},
\code{\link{getParamsFromConfig}()},
\code{\link{gitbook_medicalMonitoring_report}()},
\code{\link{knitPrintMedicalMonitoring}()},
\code{\link{render_medicalMonitoringReport}()}
}
\author{
Laure Cougnaud
}
\concept{medical monitoring reporting}
