% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-utility.R
\name{knitPrintMedicalMonitoring}
\alias{knitPrintMedicalMonitoring}
\title{Include output from medical monitoring, or list 
of such outputs in a Rmarkdown report.}
\usage{
knitPrintMedicalMonitoring(
  list,
  sep = ".",
  level = 1,
  generalLabel = "medicalMonitoring"
)
}
\arguments{
\item{list}{List of medical monitoring plots,
potentially nested.
If nested, the list should be named by the different elements,
separated by \code{sep}, e.g. 
\code{list('group1.param1' = .., 'group1.param2' = ...)}.}

\item{sep}{String with separator used to distinguish 
different levels in the labels of the list.
e.g. '.' by default.}

\item{level}{Integer with base level for section,
1 by default.}

\item{generalLabel}{String with general label for the
chunk(s) including the plots in the Rmarkdown document. 
Note: as chunks should have unique labels in a Rmd document,
this label should be unique within the same Rmarkdown document 
(in case \code{listPlots} with same names are included multiple times
in the document).}
}
\value{
No returned value, the plots are included in the
report.
}
\description{
Include output from medical monitoring, or list 
of such outputs in a Rmarkdown report.
}
\seealso{
Other medical monitoring reporting: 
\code{\link{convertMdToHtml}()},
\code{\link{getMdFromConfig}()},
\code{\link{getMdHeader}()},
\code{\link{getParamsFromConfig}()},
\code{\link{gitbook_medicalMonitoring_report}()},
\code{\link{html_medicalMonitoring_report}()},
\code{\link{render_medicalMonitoringReport}()}
}
\author{
Laure Cougnaud
}
\concept{medical monitoring reporting}
