% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-templates.R
\name{medicalMonitoring-templates}
\alias{medicalMonitoring-templates}
\title{Rmarkdown templates for medical monitoring}
\description{
Template reports with standard visualizations/tables
available in the package are described here.
}
\details{
For each template, required parameters are indicated in \strong{bold}.
}
\section{Parameter type}{Please note that the type mentioned below corresponds to the type in the config file (in YAML/JSON format).The mapping to R data type is as followed:\itemize{\item{string: }{character vector of length 1}\item{integer: }{integer vector of length 1}\item{array: }{vector/list without names}\item{object: }{list with names}}}
\item{countsVisualizationTemplate.Rmd}
\item{divisionTemplate.Rmd}
\item{listingTemplate.Rmd}
\item{patientProfilesTemplate.Rmd}
\item{plotTemplate.Rmd}
\item{summaryPlotTemplate.Rmd}
\item{summaryTableTemplate.Rmd}
