% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-templates.R
\name{medicalMonitoring-templates}
\alias{medicalMonitoring-templates}
\alias{renderTemplateDoc}
\title{Rmarkdown templates for medical monitoring}
\usage{
renderTemplateDoc()
}
\description{
Template reports with standard visualizations/tables
available in the package are described here.
}
\section{Medical monitoring template to create a report division}{
This report includes a division, i.e. extra chapter, section of subsection in the report.
The following parameters are available:
\itemize{
\item{\code{template}: }{string name of the medical monitoring template}
\item{\code{templatePackage}: }{(optional) string package from which the template should be extracted}
\item{\code{reportTitle}: }{string header title}
\item{\code{reportTitleLevel}: }{(optional) integer header level, 1 by default (1: 'chapter', '2': 'section', '3': subsection, ...)}
}
}
