% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-metadata.R
\name{renamePathDateInfoMetadata}
\alias{renamePathDateInfoMetadata}
\title{Rename variable names of metadata info}
\usage{
renamePathDateInfoMetadata(summaryInfo, namesInfo)
}
\arguments{
\item{summaryInfo}{A matrix, see output from \code{\link{getMetadata}}.}

\item{namesInfo}{Named vector to rename the final output.}
}
\value{
A matrix, same as input \code{summaryInfo} with renamed variable names.
}
\description{
Rename variable names referring to the paths and the date.
}
