% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPatientProfileTemplateReport.R
\name{runPatientProfileTemplateReport}
\alias{runPatientProfileTemplateReport}
\title{Create patient profile report, based on template.}
\usage{
runPatientProfileTemplateReport(
  dataPath,
  overwrite = FALSE,
  outputFile = "subjectProfile.pdf",
  study = "custom",
  batch = "X",
  author = NULL,
  subsetDataset = NULL,
  subsetVar = NULL,
  subsetValue = NULL,
  subsetSample = NULL,
  subjectSortDataset = NULL,
  subjectSortVar = NULL,
  subjectSortInvert = FALSE,
  exportBatchSize = 5
)
}
\arguments{
\item{dataPath}{String with path to data.}

\item{overwrite}{Logical, if TRUE (FALSE by default)
the patient profile template is overwritten in the 
output directory.}

\item{outputFile}{(optional) String with name of output file,
'subjectProfile.pdf' by default.}

\item{study}{(optional) String with study name.}

\item{batch}{(optional) String with batch name.}

\item{author}{(optional) String with author name.}

\item{subjectSortDataset}{(optional) string with dataset used to sort
the subjects for the export.}

\item{subjectSortVar}{variable(s) of \code{data} indicating the order for the subjects in the report,
by default same as \code{bookmarkVar}}

\item{subjectSortInvert}{Logical, if TRUE (FALSE by default)
subjects are sorted based on inverse order of \code{subjectSortVar}.}

\item{exportBatchSize}{Integer, number of subjects
for which patient profiles should be created at the same time,
5 by default (to optimize creation of patient profiles).}
}
\value{
Character vector with output path.
The patient profile(s) and the Rmd template
are created at \code{outputFile}.
}
\description{
Currently template is only available for SDTM data format.
}
\author{
Laure Cougnaud
}
