% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPatientProfileTemplateReport.R
\name{runPatientProfileTemplateReport}
\alias{runPatientProfileTemplateReport}
\title{Create patient profile report, based on template.}
\usage{
runPatientProfileTemplateReport(
  dataPath,
  overwrite = FALSE,
  outputFile = "subjectProfile.pdf",
  study = "custom",
  batch = "X",
  author = NULL,
  subsetDataset = NULL,
  subsetVar = NULL,
  subsetValue = NULL,
  subsetSample = NULL,
  subjectSortDataset = NULL,
  subjectSortVar = NULL,
  subjectSortDecreasing = FALSE,
  exportBatchSize = 10
)
}
\arguments{
\item{dataPath}{String with path to data.}

\item{overwrite}{Logical, if TRUE (FALSE by default)
the patient profile template is overwritten in the 
output directory.}

\item{outputFile}{(optional) String with name of output file,
'subjectProfile.pdf' by default.}

\item{study}{(optional) String with study name.}

\item{batch}{(optional) String with batch name.}

\item{author}{(optional) String with author name.}

\item{subsetDataset, subsetVar, subsetValue, subsetSample}{Dataset, variable, corresponding value and random number of subjects
to consider to subset subjects
for which patient profiles should be created.
Note that for \code{subsetSample}, a seed should
be specified upfront to get reproducible results.}

\item{subjectSortDataset, subjectSortVar, subjectSortDecreasing}{(optional) 
String with dataset name, variable name by which the subjects should
be sorted for the export, based on increasing value if \code{subjectSortDecreasing}
is FALSE (TRUE by default).}

\item{exportBatchSize}{Integer, number of subjects
for which patient profiles should be created at the same time,
10 by default (to optimize creation of patient profiles).}
}
\value{
Character vector with output path.
The patient profile(s) and the Rmd template
are created at \code{outputFile}.
}
\description{
Currently template is only available for SDTM data format.
}
\author{
Laure Cougnaud
}
