% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMonitoring.R
\name{tableMonitoring}
\alias{tableMonitoring}
\title{Create a 'monitoring table', associated to a plot.}
\usage{
tableMonitoring(
  data,
  idVar = "USUBJID",
  idLab = getLabelVar(idVar, labelVars = labelVars),
  pathVar = NULL,
  tableVars,
  tableLab = getLabelVar(tableVars, labelVars = labelVars),
  tableButton = TRUE,
  tablePars = list(),
  id = paste0("plotMonitoring", sample.int(n = 1000, size = 1)),
  labelVars = NULL
)
}
\arguments{
\item{data}{Data.frame with data.}

\item{idVar}{String with variable containing subject ID.}

\item{idLab}{String with label for \code{idVar}.}

\item{pathVar}{String with variable of \code{data} containing path
to a subject-specific report (e.g. patient profiles).
The report should be unique by element of \code{idVar}.
This report will be:
\itemize{
\item{downloaded if the user clicks on the 'p' (a.k.a 'profile') key
when hovering on a point of the plot}
\item{opened in a brower via hyperlink in the \code{idVar} of the table 
column (if specified via \code{tableVars})}
}}

\item{tableVars, tableLab}{Character vector with variables to be included 
in the table; and associated labels.}

\item{tableButton}{Logical, if TRUE (by default)
the table is included within an HTML button.}

\item{tablePars}{List with parameters passed to the
\code{\link[glpgUtilityFct]{toDTGLPG}} function.}

\item{id}{String with general id for the plot:
\itemize{
\item{'id' is used as \code{group} for the \code{\link[crosstalk]{SharedData}}}
\item{'button:[id]' is used as button ID if \code{table} is TRUE}
}
If not specified, a random id, as 'plotMonitoring[X]' is used.}

\item{labelVars}{Named character vector containing variable labels,
used by default for all labels in the plot.}
}
\value{
\code{\link[DT]{datatable}}
}
\description{
Interactive table is created, with the possibility to have
clickeable link to patient-specific report, and included
within a button.
}
\author{
Laure Cougnaud
}
